requirements.txt
streamlit
pandas
scikit-learn
joblib
plotly
bcrypt
pyjwt
openai